refactor(exporter): split internal/profile/emit.go into 3 named seams (PR-6)
Decompose the 807-LoC v2.8.0 Phase .c emission code into orchestration /
marshal / render layers. First of 4 follow-up code-health PRs after the
v2.8.0 5-PR sequence (#195-#198).

emit.go (807 → 521 LoC) keeps:
  - EmitProposals — top-level entry
  - emitOneProposal / emitTranslatedProposal — per-proposal coordinators
  - membersForProposal — orchestration helper
  - sortedTenantKeys / sortedKeys / quoteAll / joinClean / safeFilename
    — utility helpers used by every layer

emit_marshal.go (192 LoC, NEW) — YAML serialization layer:
  - marshalDefaults / marshalTenantOverride
  - withTranslatedHeader (prepend conf.d-ready translation header)
  - formatThresholdString
  - pickTenantLabelKey / tenantIDForRule / pickVaryingLabels
  - sortedStringMap (deterministic map→any for yaml.v3)
  - yamlMarshalCanonical (golden-fixture-pinned encoder)

emit_render.go (138 LoC, NEW) — Markdown rendering:
  - renderProposalMarkdown
  - renderTranslatedProposalMarkdown

Pure refactor — every function moved verbatim with imports trimmed per
file (errors / yaml dropped from emit.go since marshal owns them; sort
dropped from render since pure fmt+strings layer).

Behavioral equivalence: same 22 functions in same package — Go file
boundaries are not visible at the API surface. emit_test.go (758 LoC,
24 tests) untouched — exercises all 3 layers transparently.

Verification:
  - Linux dev container worktree path:
    `go test -race -count=1 ./...` — 9/9 packages pass
  - internal/profile coverage unchanged (91.5% per nightly bench-record)
  - Same Pattern as PR-3 (god-method → named seams) so reviewers
    have an established mental model

Sets v2.8.0 Phase .c new code up with the same code-health discipline
the rest of the exporter got from PR-1 through PR-5.
