+++ components/threshold-exporter/app/internal/profile/emit.go
@@ -48,13 +48,10 @@ package profile
 // InMemoryConfigSource and C-10 batchpr.Plan: keep IO at the edges.
 
 import (
-	"errors"
 	"fmt"
 	"path"
 	"sort"
 	"strings"
-
-	"gopkg.in/yaml.v3"
 
 	"github.com/vencil/threshold-exporter/internal/parser"
 )
@@ -453,283 +450,6 @@ func emitTranslatedProposal(
 // _defaults.yaml body so reviewers see provenance + warnings
 // directly in the file (helpful when reviewing a Git diff without
 // PROPOSAL.md context).
-func withTranslatedHeader(body []byte, prop ExtractionProposal, translation *ProposalTranslation) []byte {
-	header := strings.Builder{}
-	header.WriteString("# Generated by C-9 Profile Builder PR-3 (PromRule → conf.d translator).\n")
-	header.WriteString(fmt.Sprintf("# Cluster: %d member rules; dialect=%s; confidence=%s; translation=%s.\n",
-		len(prop.MemberRuleIDs), prop.Dialect, prop.Confidence, translation.Status))
-	if translation.Operator != "" {
-		header.WriteString(fmt.Sprintf("# Comparison operator: `%s` (recorded for ADR-019 §emit-direction follow-ups).\n",
-			translation.Operator))
-	}
-	if translation.Severity != "" {
-		header.WriteString(fmt.Sprintf("# Severity: `%s`.\n", translation.Severity))
-	}
-	for _, w := range translation.Warnings {
-		header.WriteString("# WARN: " + w + "\n")
-	}
-	header.WriteString("# See PROPOSAL.md for member listing + per-tenant overrides.\n\n")
-	return append([]byte(header.String()), body...)
-}
-
-// formatThresholdString renders a float64 threshold value into the
-// quoted-string form ResolveAt expects in tenant overrides. Drops
-// trailing zeros so common values like 80.0 emit as "80".
-func formatThresholdString(v float64) string {
-	if v == float64(int64(v)) {
-		return fmt.Sprintf("%d", int64(v))
-	}
-	return strings.TrimRight(strings.TrimRight(fmt.Sprintf("%f", v), "0"), ".")
-}
-
-// marshalDefaults builds the YAML content for `_defaults.yaml`.
-// The shape carries the proposal-level shared structure plus
-// metadata (dialect, provenance) that PR-3's translator will
-// consume to produce the final conf.d/-ready dict.
-func marshalDefaults(prop ExtractionProposal) ([]byte, error) {
-	doc := map[string]any{
-		"shared_expr_template": prop.SharedExprTemplate,
-		"dialect":              prop.Dialect,
-		"prom_portable":        prop.Dialect == string(parser.DialectProm),
-		"member_count":         len(prop.MemberRuleIDs),
-		"confidence":           string(prop.Confidence),
-	}
-	if prop.SharedFor != "" {
-		doc["shared_for"] = prop.SharedFor
-	}
-	if len(prop.SharedLabels) > 0 {
-		doc["shared_labels"] = sortedStringMap(prop.SharedLabels)
-	}
-	if len(prop.VaryingLabelKeys) > 0 {
-		doc["varying_label_keys"] = append([]string(nil), prop.VaryingLabelKeys...)
-	}
-	return yamlMarshalCanonical(doc)
-}
-
-// marshalTenantOverride builds the per-tenant file. Wraps under
-// `tenants:` so the existing extractTenantRaw() in
-// config_inheritance.go can parse it (forward-compatibility — once
-// the translator lands, this shape becomes the actual tenant.yaml).
-func marshalTenantOverride(tenantID string, rule parser.ParsedRule, prop ExtractionProposal) ([]byte, error) {
-	tenantBlock := map[string]any{
-		"source_rule_id": rule.SourceRuleID,
-		"expr":           rule.Expr,
-	}
-	if rule.Alert != "" {
-		tenantBlock["alert"] = rule.Alert
-	}
-	if rule.Record != "" {
-		tenantBlock["record"] = rule.Record
-	}
-	if rule.For != "" && rule.For != prop.SharedFor {
-		tenantBlock["for"] = rule.For
-	}
-	// Only emit the labels that vary across the cluster — the
-	// shared ones live in _defaults.yaml.
-	varyingLabels := pickVaryingLabels(rule.Labels, prop)
-	if len(varyingLabels) > 0 {
-		tenantBlock["labels"] = sortedStringMap(varyingLabels)
-	}
-	if len(rule.Annotations) > 0 {
-		// Annotations don't participate in clustering (they're
-		// human text); keep them per-tenant verbatim.
-		tenantBlock["annotations"] = sortedStringMap(rule.Annotations)
-	}
-
-	doc := map[string]any{
-		"tenants": map[string]any{
-			tenantID: tenantBlock,
-		},
-	}
-	return yamlMarshalCanonical(doc)
-}
-
-// pickTenantLabelKey heuristically picks the label key that
-// distinguishes members of a proposal — typically `tenant`, but
-// the cluster may use `team` or other axes. Returns "" when no
-// varying labels exist (caller falls back to a synthetic id).
-//
-// PR-2 heuristic: prefer "tenant" if present in VaryingLabelKeys,
-// else pick the first VaryingLabelKey alphabetically. Stable.
-func pickTenantLabelKey(prop ExtractionProposal) string {
-	for _, k := range prop.VaryingLabelKeys {
-		if k == "tenant" {
-			return "tenant"
-		}
-	}
-	if len(prop.VaryingLabelKeys) > 0 {
-		// VaryingLabelKeys is already sorted (PR-1 contract).
-		return prop.VaryingLabelKeys[0]
-	}
-	return ""
-}
-
-// tenantIDForRule extracts the tenant identifier from a rule's
-// Labels using the picked tenant key. Falls back to "" when the
-// rule has no value for that key.
-func tenantIDForRule(rule parser.ParsedRule, tenantKey string) string {
-	if tenantKey == "" {
-		return ""
-	}
-	v, ok := rule.Labels[tenantKey]
-	if !ok {
-		return ""
-	}
-	return v
-}
-
-// pickVaryingLabels returns the subset of `labels` whose keys are
-// in VaryingLabelKeys. The returned map is freshly allocated.
-func pickVaryingLabels(labels map[string]string, prop ExtractionProposal) map[string]string {
-	if len(labels) == 0 || len(prop.VaryingLabelKeys) == 0 {
-		return nil
-	}
-	varyingSet := make(map[string]struct{}, len(prop.VaryingLabelKeys))
-	for _, k := range prop.VaryingLabelKeys {
-		varyingSet[k] = struct{}{}
-	}
-	out := make(map[string]string)
-	for k, v := range labels {
-		if _, ok := varyingSet[k]; ok {
-			out[k] = v
-		}
-	}
-	return out
-}
-
-// sortedStringMap converts a map[string]string into a
-// yaml.MapSlice-equivalent representation. Returning a plain
-// map[string]any is enough — yaml.v3 sorts keys alphabetically by
-// default for map[string]any, so output is deterministic.
-func sortedStringMap(in map[string]string) map[string]any {
-	out := make(map[string]any, len(in))
-	for k, v := range in {
-		out[k] = v
-	}
-	return out
-}
-
-// renderProposalMarkdown produces the PROPOSAL.md body — a human
-// summary so a reviewer can decide accept / reject without diving
-// into the YAML files.
-func renderProposalMarkdown(propIdx int, prop ExtractionProposal, tenantKey string) string {
-	out := strings.Builder{}
-	out.WriteString(fmt.Sprintf("# Proposal %d\n\n", propIdx))
-	out.WriteString(fmt.Sprintf("**Confidence**: %s  \n", prop.Confidence))
-	out.WriteString(fmt.Sprintf("**Dialect**: %s  \n", prop.Dialect))
-	out.WriteString(fmt.Sprintf("**Members**: %d rules  \n", len(prop.MemberRuleIDs)))
-	if tenantKey != "" {
-		out.WriteString(fmt.Sprintf("**Tenant key**: `%s`  \n", tenantKey))
-	} else {
-		out.WriteString("**Tenant key**: _(none — proposal has no varying labels)_  \n")
-	}
-	out.WriteString(fmt.Sprintf("**Estimated YAML lines saved**: %d  \n\n", prop.EstimatedYAMLLineSavings))
-
-	out.WriteString("## Reason\n\n")
-	out.WriteString(prop.Reason)
-	if !strings.HasSuffix(prop.Reason, "\n") {
-		out.WriteString("\n")
-	}
-	out.WriteString("\n")
-
-	out.WriteString("## Member rules\n\n")
-	for _, rid := range prop.MemberRuleIDs {
-		out.WriteString(fmt.Sprintf("- `%s`\n", rid))
-	}
-	out.WriteString("\n")
-
-	out.WriteString("## Shared structure\n\n")
-	out.WriteString(fmt.Sprintf("- Expression template (normalised): `%s`\n", prop.SharedExprTemplate))
-	if prop.SharedFor != "" {
-		out.WriteString(fmt.Sprintf("- For: `%s`\n", prop.SharedFor))
-	}
-	if len(prop.SharedLabels) > 0 {
-		out.WriteString("- Shared labels:\n")
-		for _, k := range sortedKeys(prop.SharedLabels) {
-			out.WriteString(fmt.Sprintf("  - `%s`: `%s`\n", k, prop.SharedLabels[k]))
-		}
-	}
-	if len(prop.VaryingLabelKeys) > 0 {
-		out.WriteString(fmt.Sprintf("- Varying label keys: %s\n", strings.Join(quoteAll(prop.VaryingLabelKeys), ", ")))
-	}
-
-	out.WriteString("\n---\n\n")
-	out.WriteString("_PR-2 emission: this artifact tree is intermediate. PR-3 will translate it into the final ADR-018 conf.d/ shape._\n")
-	return out.String()
-}
-
-// renderTranslatedProposalMarkdown is the PR-3 sibling of
-// renderProposalMarkdown — same overall shape, but adds a
-// "Translation summary" section that surfaces metric_key, default
-// threshold, severity, and per-tenant overrides so a reviewer can
-// see the conf.d-shape decisions without opening the YAML files.
-func renderTranslatedProposalMarkdown(propIdx int, prop ExtractionProposal, translation *ProposalTranslation, tenantKey string) string {
-	out := strings.Builder{}
-	out.WriteString(fmt.Sprintf("# Proposal %d (translated to conf.d)\n\n", propIdx))
-	out.WriteString(fmt.Sprintf("**Confidence**: %s  \n", prop.Confidence))
-	out.WriteString(fmt.Sprintf("**Dialect**: %s  \n", prop.Dialect))
-	out.WriteString(fmt.Sprintf("**Translation status**: %s  \n", translation.Status))
-	out.WriteString(fmt.Sprintf("**Members**: %d rules  \n", len(prop.MemberRuleIDs)))
-	if tenantKey != "" {
-		out.WriteString(fmt.Sprintf("**Tenant key**: `%s`  \n", tenantKey))
-	}
-	out.WriteString("\n")
-
-	out.WriteString("## Translation summary\n\n")
-	out.WriteString(fmt.Sprintf("- Metric key: `%s`\n", translation.MetricKey))
-	out.WriteString(fmt.Sprintf("- Default threshold (cluster median): `%g`\n", translation.DefaultThreshold))
-	if translation.Operator != "" {
-		out.WriteString(fmt.Sprintf("- Comparison operator: `%s`\n", translation.Operator))
-	}
-	if translation.Severity != "" {
-		out.WriteString(fmt.Sprintf("- Severity: `%s`\n", translation.Severity))
-	}
-	if len(translation.PerTenantOverrides) > 0 {
-		out.WriteString("- Per-tenant overrides (only tenants whose value differs from default):\n")
-		for _, k := range sortedTenantKeys(translation.PerTenantOverrides) {
-			out.WriteString(fmt.Sprintf("  - `%s` → `%g`\n", k, translation.PerTenantOverrides[k]))
-		}
-	}
-	out.WriteString("\n")
-
-	if len(translation.Warnings) > 0 {
-		out.WriteString("## Translator warnings\n\n")
-		for _, w := range translation.Warnings {
-			out.WriteString("- " + w + "\n")
-		}
-		out.WriteString("\n")
-	}
-
-	out.WriteString("## Member rules\n\n")
-	for _, m := range translation.MemberStatuses {
-		switch m.Status {
-		case TranslationOK:
-			out.WriteString(fmt.Sprintf("- ✅ `%s` — threshold=`%g`\n", m.SourceRuleID, m.Threshold))
-		case TranslationPartial:
-			out.WriteString(fmt.Sprintf("- ⚠️ `%s` — threshold=`%g` (partial: %s)\n", m.SourceRuleID, m.Threshold, strings.Join(m.Warnings, "; ")))
-		case TranslationSkipped:
-			out.WriteString(fmt.Sprintf("- ⏭️ `%s` — skipped (%s)\n", m.SourceRuleID, m.SkipReason))
-		}
-	}
-	out.WriteString("\n")
-
-	out.WriteString("## Reason\n\n")
-	out.WriteString(prop.Reason)
-	if !strings.HasSuffix(prop.Reason, "\n") {
-		out.WriteString("\n")
-	}
-	out.WriteString("\n")
-
-	out.WriteString("---\n\n")
-	out.WriteString("_PR-3 emission: this proposal is conf.d-ready (deepMerge-compatible). Review_  \n")
-	out.WriteString("_`_defaults.yaml` and per-tenant `<id>.yaml` then commit; the threshold-exporter_  \n")
-	out.WriteString("_runtime ResolveAt path consumes them via ADR-018 inheritance._\n")
-	return out.String()
-}
-
-// sortedTenantKeys returns map keys sorted alphabetically (helper
-// for renderTranslatedProposalMarkdown so the per-tenant override
-// list is deterministic).
 func sortedTenantKeys(m map[string]float64) []string {
 	out := make([]string, 0, len(m))
 	for k := range m {
@@ -799,9 +519,3 @@ func safeFilename(s string) string {
 //   - centralise the marshal call (single place to add header
 //     comments later if PR-3 wants a `# generated by ...` line).
 //   - turn yaml errors into a sentinel for the warnings collector.
-func yamlMarshalCanonical(doc map[string]any) ([]byte, error) {
-	if doc == nil {
-		return nil, errors.New("yaml marshal: nil doc")
-	}
-	return yaml.Marshal(doc)
-}
