+++ components/threshold-exporter/app/internal/profile/emit_marshal.go
@@ -1,0 +1,192 @@
+package profile
+
+// YAML serialization layer for ProposalSet / ExtractionProposal artifacts.
+// Split out of emit.go in v2.8.0 PR-6.
+//
+// Concerns covered here:
+//   - marshalDefaults / marshalTenantOverride — top-level YAML emit
+//   - withTranslatedHeader — prepend conf.d-ready translation header
+//   - formatThresholdString — float→string with stable precision
+//   - pickTenantLabelKey / tenantIDForRule / pickVaryingLabels —
+//     tenant-axis extraction from ParsedRule labels
+//   - sortedStringMap — deterministic map→any ordering for yaml.v3
+//   - yamlMarshalCanonical — golden-fixture-pinned canonical encoder
+//
+// Pure functions; no IO, no goroutines. The caller (emit.go orchestration)
+// composes them into the per-proposal artifact tree.
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+
+	"gopkg.in/yaml.v3"
+
+	"github.com/vencil/threshold-exporter/internal/parser"
+)
+
+func withTranslatedHeader(body []byte, prop ExtractionProposal, translation *ProposalTranslation) []byte {
+	header := strings.Builder{}
+	header.WriteString("# Generated by C-9 Profile Builder PR-3 (PromRule → conf.d translator).\n")
+	header.WriteString(fmt.Sprintf("# Cluster: %d member rules; dialect=%s; confidence=%s; translation=%s.\n",
+		len(prop.MemberRuleIDs), prop.Dialect, prop.Confidence, translation.Status))
+	if translation.Operator != "" {
+		header.WriteString(fmt.Sprintf("# Comparison operator: `%s` (recorded for ADR-019 §emit-direction follow-ups).\n",
+			translation.Operator))
+	}
+	if translation.Severity != "" {
+		header.WriteString(fmt.Sprintf("# Severity: `%s`.\n", translation.Severity))
+	}
+	for _, w := range translation.Warnings {
+		header.WriteString("# WARN: " + w + "\n")
+	}
+	header.WriteString("# See PROPOSAL.md for member listing + per-tenant overrides.\n\n")
+	return append([]byte(header.String()), body...)
+}
+
+// formatThresholdString renders a float64 threshold value into the
+// quoted-string form ResolveAt expects in tenant overrides. Drops
+// trailing zeros so common values like 80.0 emit as "80".
+func formatThresholdString(v float64) string {
+	if v == float64(int64(v)) {
+		return fmt.Sprintf("%d", int64(v))
+	}
+	return strings.TrimRight(strings.TrimRight(fmt.Sprintf("%f", v), "0"), ".")
+}
+
+// marshalDefaults builds the YAML content for `_defaults.yaml`.
+// The shape carries the proposal-level shared structure plus
+// metadata (dialect, provenance) that PR-3's translator will
+// consume to produce the final conf.d/-ready dict.
+func marshalDefaults(prop ExtractionProposal) ([]byte, error) {
+	doc := map[string]any{
+		"shared_expr_template": prop.SharedExprTemplate,
+		"dialect":              prop.Dialect,
+		"prom_portable":        prop.Dialect == string(parser.DialectProm),
+		"member_count":         len(prop.MemberRuleIDs),
+		"confidence":           string(prop.Confidence),
+	}
+	if prop.SharedFor != "" {
+		doc["shared_for"] = prop.SharedFor
+	}
+	if len(prop.SharedLabels) > 0 {
+		doc["shared_labels"] = sortedStringMap(prop.SharedLabels)
+	}
+	if len(prop.VaryingLabelKeys) > 0 {
+		doc["varying_label_keys"] = append([]string(nil), prop.VaryingLabelKeys...)
+	}
+	return yamlMarshalCanonical(doc)
+}
+
+// marshalTenantOverride builds the per-tenant file. Wraps under
+// `tenants:` so the existing extractTenantRaw() in
+// config_inheritance.go can parse it (forward-compatibility — once
+// the translator lands, this shape becomes the actual tenant.yaml).
+func marshalTenantOverride(tenantID string, rule parser.ParsedRule, prop ExtractionProposal) ([]byte, error) {
+	tenantBlock := map[string]any{
+		"source_rule_id": rule.SourceRuleID,
+		"expr":           rule.Expr,
+	}
+	if rule.Alert != "" {
+		tenantBlock["alert"] = rule.Alert
+	}
+	if rule.Record != "" {
+		tenantBlock["record"] = rule.Record
+	}
+	if rule.For != "" && rule.For != prop.SharedFor {
+		tenantBlock["for"] = rule.For
+	}
+	// Only emit the labels that vary across the cluster — the
+	// shared ones live in _defaults.yaml.
+	varyingLabels := pickVaryingLabels(rule.Labels, prop)
+	if len(varyingLabels) > 0 {
+		tenantBlock["labels"] = sortedStringMap(varyingLabels)
+	}
+	if len(rule.Annotations) > 0 {
+		// Annotations don't participate in clustering (they're
+		// human text); keep them per-tenant verbatim.
+		tenantBlock["annotations"] = sortedStringMap(rule.Annotations)
+	}
+
+	doc := map[string]any{
+		"tenants": map[string]any{
+			tenantID: tenantBlock,
+		},
+	}
+	return yamlMarshalCanonical(doc)
+}
+
+// pickTenantLabelKey heuristically picks the label key that
+// distinguishes members of a proposal — typically `tenant`, but
+// the cluster may use `team` or other axes. Returns "" when no
+// varying labels exist (caller falls back to a synthetic id).
+//
+// PR-2 heuristic: prefer "tenant" if present in VaryingLabelKeys,
+// else pick the first VaryingLabelKey alphabetically. Stable.
+func pickTenantLabelKey(prop ExtractionProposal) string {
+	for _, k := range prop.VaryingLabelKeys {
+		if k == "tenant" {
+			return "tenant"
+		}
+	}
+	if len(prop.VaryingLabelKeys) > 0 {
+		// VaryingLabelKeys is already sorted (PR-1 contract).
+		return prop.VaryingLabelKeys[0]
+	}
+	return ""
+}
+
+// tenantIDForRule extracts the tenant identifier from a rule's
+// Labels using the picked tenant key. Falls back to "" when the
+// rule has no value for that key.
+func tenantIDForRule(rule parser.ParsedRule, tenantKey string) string {
+	if tenantKey == "" {
+		return ""
+	}
+	v, ok := rule.Labels[tenantKey]
+	if !ok {
+		return ""
+	}
+	return v
+}
+
+// pickVaryingLabels returns the subset of `labels` whose keys are
+// in VaryingLabelKeys. The returned map is freshly allocated.
+func pickVaryingLabels(labels map[string]string, prop ExtractionProposal) map[string]string {
+	if len(labels) == 0 || len(prop.VaryingLabelKeys) == 0 {
+		return nil
+	}
+	varyingSet := make(map[string]struct{}, len(prop.VaryingLabelKeys))
+	for _, k := range prop.VaryingLabelKeys {
+		varyingSet[k] = struct{}{}
+	}
+	out := make(map[string]string)
+	for k, v := range labels {
+		if _, ok := varyingSet[k]; ok {
+			out[k] = v
+		}
+	}
+	return out
+}
+
+// sortedStringMap converts a map[string]string into a
+// yaml.MapSlice-equivalent representation. Returning a plain
+// map[string]any is enough — yaml.v3 sorts keys alphabetically by
+// default for map[string]any, so output is deterministic.
+func sortedStringMap(in map[string]string) map[string]any {
+	out := make(map[string]any, len(in))
+	for k, v := range in {
+		out[k] = v
+	}
+	return out
+}
+
+// renderProposalMarkdown produces the PROPOSAL.md body — a human
+// summary so a reviewer can decide accept / reject without diving
+// into the YAML files.
+func yamlMarshalCanonical(doc map[string]any) ([]byte, error) {
+	if doc == nil {
+		return nil, errors.New("yaml marshal: nil doc")
+	}
+	return yaml.Marshal(doc)
+}
